+++ OUDS/Core/Tokens/RawTokens/Sources/Values/GridRawTokens+Values.swift
@@ -2,29 +2,18 @@
 // Software Name: OUDS iOS
 // SPDX-FileCopyrightText: Copyright (c) Orange SA
 // SPDX-License-Identifier: MIT
-// 
+//
 // This software is distributed under the MIT license,
 // the text of which is available at https://opensource.org/license/MIT/
 // or see the "LICENSE" file for more details.
-// 
+//
 // Authors: See CONTRIBUTORS.txt
-// Software description: A SwiftUI components library with code examples for Orange Unified Design System 
+// Software description: A SwiftUI components library with code examples for Orange Unified Design System
 //
 
-import Foundation
-
-// MARK: - Type aliases to keep grammar clear
-
-/// In the global design system tool, the notion of "raw tokens of grids" exist but refers always to dimension raw tokens.
-public typealias GridRawToken = DimensionRawToken
-
-// MARK: - Raw tokens
-
-/// This is the group of all **raw tokens** related to **grid**.
-/// Primitive types such as `Int` must be used to as to allow to use `@objc` keywords in extensions for overriding.
-/// Such tokens are packed in a _Swift enum_ so as to gather them in one object with the suitable namespace and avoid to have just constants in nothing else
-///  (i.e. publicly accessible from everywhere). More optimized than _struct_.
-public enum GridRawTokens {
+/// Extracted in a separated file to help the *Figma* JSON to Swift parser to generate files to include easily.
+/// Should be fully generated in the future.
+extension GridRawTokens {
 
     private static let dimensionBase: GridRawToken = 4